pyproject.toml
[tool.poetry]
name = "billing"
version = "20.0.0"
description = "Toy billing system"
authors = ["Artem.Angelchev"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
python-status = "^1.0.1"
pydantic = "^1.6.1"
flask = "^1.1.2"
flask-compress = "^1.5.0"
flask-sqlalchemy = "^2.4.4"
flask-migrate = "^2.5.3"
psycopg2-binary = "^2.8.5"
loguru = "^0.5.1"
gunicorn = "^20.0.4"

pytest = {version = "^5.2", optional = true}
pytest-mock = {version = "^1.11", optional = true}
pytest-cov = {version = "^2.8", optional = true}

[tool.poetry.dev-dependencies]
pylint = "^2.5.3"
flake8 = "^3.8.3"

[tool.poetry.extras]
test = [
  "pytest",
  "pytest-cov",
  "pytest-mock",
]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
